Merge term bindings from evil-special-modes
Previous, we had evil-ansi-term here.  I think the name evil-term is more
appropriate as it supports term, ansi-term and multi-term.

I've merged the hook that synchronizes normal/line mode and insert/char mode.
Because of this, there is no need to add bindings to send raw characters in
insert mode.

I've integrated the feature of not moving the cursor back: the new
implementation does not rebind <escape>, it sets `evil-move-cursor-back locally.