Add support for dd in wgrep buffers
This seems to be the only way to define dd without losing the functionality of
the rest of the `d`-prefixed keys. This should allow regular `d` operations as
well as `dd` to remove the target line.

This is just a copy of the `evil-delete` operator from evil master source. And
builds upon what I mentioned earlier in:

https://github.com/emacs-evil/evil-collection/pull/111