neotree: Replace neotree-make-executor with custom macro
neotree-make-executor is a macro, so it will get expanded correctly at compile
time if neotree is available. If neotree is not available, somehow
neotree-make-executor gets tagged as an invalid function at compile time, even
if neotree is available at run time.

It's not a good idea to call a macro from a foreign library when you don't know
if that library will be available at compile time. There are other alternatives
to this fix, but since neotree-make-executor is so simple this seems like the
best option.

Fixes #158