Fix loading of evil-collection-X from subdirs
Create autoloads for each of the setup functions in evil-collection. Add
evil-collection-require, which is a version of require that knows where the
evil-collection files live.

Use evil-collection-require to fix header in evil-collection-woman.

Update readme to reflect the fact that it's no longer necessary to load the
evil-collection-X file to run evil-collection-X-setup