Magit: Properly format rebase instructions when commentChar is "%"
This change adjusts the formatting of the instructions in the `git-rebase`
buffer to allow them to be properly formatted when git's `core.commentChar` is
set to `%`.  In particular, lines will look something like this:

```
%-8s p
```

The root cause seems to be the how the lines are constructed: comment characters
are directly added to the format string.  When it is `%`, `format` appears to
proceed until it finds the next `%` character, (corresponding to the `%` in
`%-8s`) above and prints it directly to the string rather than interpreting it
something to format.

This is fixed by constructing the format string with the comment character as an
argument to `format` rather than being added directly to the format string.

Fixes #468.