vterm: make the `G' button jump to the prompt line
AFAIK `vterm` is basically a square of the size of the whole window it
is placed in. As a consequence `M-x count-words` shows 30 lines even
for freshly spawned `vterm` with just one line of text. And as a
consequence `G` always jumps down to the bottom of the window, even
though it is an unused line.

This patch changes the behavior and makes `G` jump to the last prompt
line. Which in ideal world, should be the last line of the buffer.