Use evil-lookup-key when translating keys
Makes it possible to indiscriminately translate multi-key sequences
like "gj" without errors. Using plain `lookup-key` will end up binding
a bunch of keys to a number like 1 instead of a command, because it
returns a number when it's passed an invalid sequence of prefix
characters. `evil-lookup-key` discards those numbers and returns nil
instead.