Only load mu4e-compatibility code for mu4e 1.8 or older, not 1.9
The 1.9 release numbers were the pre-releases where the new features and the new
interface were introduced. They were not packaged in distro repositories,
therefore I would not load backwards compatibility code for that. 1.8 is stable
and packaged in many linux distro repositories.

Further, this improved compatability with the file name that were are loading.