fix prefix key collision for mu4e
This reverts commit 05731c551be8cdda40ae6479adfb30b7e9c7fe39.

This fixes an issue that arises when loading `evil-collection-mu4e`, which
gives:

```
Key sequence c c starts with non-prefix key c
```

This occurs for both `mu4e-headers-mode-map` and `mu4e-main-mode-map`.

This should probably just be a different key, but for now I'm just trying to get
it unbroken.

cc @RyanGibb for visibility.

Thanks for the wonderful library!